chore: remove wit_parser glob imports
This was causing confusion on certain rust-analyzer versions as it
confused types and traits imported by the `wasm_parser::*` in lib.rs, and our own types which were exported as `pub use crate::{mod}::*`. Since each module is a parent of lib.rs, both are visible, using rustc, the local bindings take precedence, while rust-analyzer appears to favor the first in order (wit_parser).

This lead to false-positive error, incorrect completion and types, and broken `goto-defintion`, as rust-analyzer took you to the wit-parser source code.